FIX: Parallelism failure in IIfcIndexedPolyCurve
I have a few models that fail with a memory violation exception when
building wires for IIfcIndexedPolyCurve.
The error occurs in the EXEC_NATIVE->BuildWire call and results in
a crash, but only when executed in multithreading.

EXEC_NATIVE->BuildWire is not thread safe, so I've added a lock guard
to ensure that each call waits for the completion of the previous and
the crash is prevented.

Other changes are introduced to improve the trace logging to investigate
unexpected crashes in win x64 environments on third party comptuters
that are difficult to reproduce.

The regression exe has been changed to allow for engine selection
configuration.

FactoryBase.h has been changed to provide more consistent logging
messages across configurations.